Make Pipeline settings generation reusable for consuming projects.
We use a fairly custom Pipeline configuration for Djblets static media,
making use of wrappers for the LessCSS compiler and a standard set of
options and plugins for Babel and LessCSS. Consuming applications
(namely Review Board) use the same options, and have previously had to
duplicate the configuration. This meant that they didn't gain any
benefits from improvements made to this configuration.

This change introduces a new `djblets.pipeline.settings` module, which
contains a `build_pipeline_settings()` function that builds a standard
`PIPELINE` settings dictionary. It provides the standard settings we
use, but allows a degree of customization. Consumers can make use of
this directly, or modify the generated settings after the fact.

Testing Done:
Built a Djblets package successfully.

Unit tests pass.

Made use of this in Review Board and verified the resulting static media.

Reviewed at https://reviews.reviewboard.org/r/11367/