Update Djblets 1.0.x to build with modern Node.JS.
Djblets 2.1 got support for building with a modern Node.JS, largely
thanks to our fork of the LessCSS autoprefix plugin. This change is
simpler in scope, but makes use of that plugin and disables some strict
argument handling in `less-imports.js` to enable it to run. This allows
us to build Djblets packages on modern Node.

Testing Done:
Successfully built a package on Node.JS 14.

Reviewed at https://reviews.reviewboard.org/r/11386/