Drop support for Django 1.6 for Djblets 3.0.
Djblets 3.0 is a transition release, dropping Django 1.6 support but
retaining Django 1.11 support and adding in Django 3.2 TLS support.
We'll be moving Review Board 4.0 to this in time, and then finally
completely dropping older non-LTS releases of Django and Python 2.x in
Djblets 4.0.

To start, we're dropping Django 1.6 support, leaving only 1.11 as a
compatible version. We won't be adding 3.2 support until the codebase
is ready to work with it.

Testing Done:
Ran `setup.py develop` in a virtualenv with 1.6. Saw that it installed
Django 1.11.

Reviewed at https://reviews.reviewboard.org/r/11619/