Switch URL-related imports to using django.urls.
Old versions of Django used `django.core.urlresolvers` as the module for
most URL-related imports. That's since been changed to live in
`django.urls`. Along with this, some of the names have changed (all
`Regex` prefixes were dropped, for instance).

This change updates all of our imports to use the new imports and names.
This is compatible with both Django 1.11 and 3.2.

Testing Done:
Unit tests pass on Django 1.11 and on 3.2 (along with other in-development
changes).

Reviewed at https://reviews.reviewboard.org/r/11622/