Remove providing_args from all Signal instances.
`Signal` has historically contained a `providing_args=` argument, which
listed the names of the arguments expected to be sent to the signal.
This was optional and never used for anything beyond the ability to
introspect a signal (which probably nobody does), and isn't supported on
modern versions of Django.

This change removes `providing_args=` from all signals.

Testing Done:
Unit tests pass on Django 1.11 and on 3.2 (with other in-progress changes).

Reviewed at https://reviews.reviewboard.org/r/11623/