Drop support for Python 2.7 and 3.6.
Djblets 3 drops support for Python 2.7 and 3.6. Python 2.7 reached
end-of-life, and 3.6 will reach end-of-life on December 23, 2021.

Python 3.7 is now the minimum supported version. It reaches end-of-life
on June 27, 2023.

Both 2.7 and 3.6 have been removed from the classifiers, dependency
ranges, and minimum versions. The old `PYTHON_2_*` constants have been
removed as well. This is not considered an API change, as the module
explicitly states that the contents within may change.

The package is no longer built as a universal wheel, and `setup.py`
won't run on older versions of Python.

Testing Done:
Unit tests pass.

Built a package. Verified it was a Python 3-only wheel with the correct
metadata.

Reviewed at https://reviews.reviewboard.org/r/11833/