Update all scripts to use python3 instead of python.
This updates all scripts in the Djblets tree to use `python3` instead of
`python`, ensuring that the right version is used when in an environment
containing both versions.

`run-pyflakes.py` needed some changes to run on Python 3. While we may
want to remove this script, that's a decision that'll be left for after
this update.

Testing Done:
Tested all scripts and made sure they ran.

Reviewed at https://reviews.reviewboard.org/r/11834/