Update URL patterns to use new path() and re_path().
Django 2+ provides a new method for definings URLs, `path()`. This is
much simpler for basic cases, and makes URL patterns more readable. In
addition, the old `url()` method has been renamed to `re_path()`, and a
warning is raised when using the old name, so switch to that too.

Testing Done:
Ran unit tests.