Ignore a specific six-related import warning on Python 3.10 in tests.
The version of `six` in all versions of Django can trigger the following
`ImportWarning` on Python 3.10:

    _SixMetaPathImporter.find_spec() not found; falling back to
    find_module()

This wasn't an issue with the old test runner, but is being hit in
`pytest`, causing tests to fail.

The simplest solution, unfortunately, is just to filter this out in
`assertWarns`. This implements that filter, fixing those tests.

Testing Done:
Unit tests pass on all supported versions of Python.

Reviewed at https://reviews.reviewboard.org/r/11930/