Invoke session middleware correctly.
For a while, middleware was allowed to be invoked in the old way
(instantiate with no arguments and then call `process_request`). This is
going away, and we'll need to instantiate with a `get_response` callable
and go through the new-style chain. This change updates our tests which
manually invoke middleware to do that.

Testing Done:
Ran unit tests.

Reviewed at https://reviews.reviewboard.org/r/11977/