Update pipeline storage class.
The old `PipelineCachedStorage` class has been removed, because Django
removed the `CachedStaticFilesStorage` class. The reasoning here is that
this storage backend had non-specified "intractable problems". Django
advises to switch to `ManifestStaticFilesStorage`, and Pipeline does
have a storage class that inherits from this. This class does the same
hashing for filenames that the old cached storage does. The major
difference is that it also looks for those files in the manifest file.
This can be disabled via a class attribute, so I've added a subclass of
it that uses the backend for its hashing but turns off manifest
checking.

Testing Done:
Ran unit tests.

Reviewed at https://reviews.reviewboard.org/r/11979/