Add flake8/pep8 settings to setup.cfg
Review Bot will run flake8 when changes are posted for review, but it's
useful to be able to use it locally. This change adds the relevant
configuration so that it's run with the same settings, to prevent
warnings about things we don't care about.

Testing Done:
Ran flake8 and saw it ignore some things we don't use.

Reviewed at https://reviews.reviewboard.org/r/11959/