Fix errors in conditions JS.
The version of Backbone we're using apparently doesn't like passing a
fat arrow function for the `model` constructor attribute of a new
collection. This was causing various errors to pop up when attempting to
use the conditions UI (and failures for those code paths in the
js-tests). This change updates it to use a regular `function` instead.

Testing Done:
- Ran js-tests.
- Was able to configure a condition set for an integration in Review
  Board.

Reviewed at https://reviews.reviewboard.org/r/12298/