Add a new condition value field for multiple-choice selection.
We have existing code that makes it easy to add multiple-choice
selections for models based on a queryset, but we didn't have anything
for multiple-choice selections based on just a set of choices. This
change adds a value field to do that.

Testing Done:
Used this in Review Board. Saw that existing values were loaded properly
and new values were saved correctly.

Reviewed at https://reviews.reviewboard.org/r/12330/