Update the extension failure error to use the modern button text.
When an extension fails to load, we give basic instructions on how to
fix it. Those instructions reference "Scan for installed extensions",
which is an old label for the button.

This change updates the instructions to say "Reload Extensions" instead.

Testing Done:
Loaded the extensions list, which had a bunch of missing extensions. Saw
the revised text.

Reviewed at https://reviews.reviewboard.org/r/12401/