Update thumbnail templatetag to use non-deprecated name for antialising.
The templatetag we use for creating image thumbnails was using a
long-deprecated name for the antialias flag. This was causing a
`DeprecationWarning` when running tests. This change updates us to use
the new name.

Testing Done:
Ran Review Board unit tests and saw that the deprecation warning was
gone.

Reviewed at https://reviews.reviewboard.org/r/12406/