Fix a regression in cache_memoize with Python 3.7.
The recent work on the `cache_memoize()` logic introduced some syntax
that's only compatible with Python 3.8. The code in question tried to
use the `yield a, *b` syntax to yield a tuple composed of `a` and the
contents of tuple `b`. This does not work in Python 3.7.

Instead, we're now using `yield (a,) + b`, which is the equivalent
behavior.

Testing Done:
All unit tests pass on Python 3.7 through 3.10.

Reviewed at https://reviews.reviewboard.org/r/12427/