Fix a crash that could occur when patching the collections module.
We attempt to patch the `collections` module when on Python 3.10, in
order to provide some backwards-compatibility imports. It does a plain
`import collections`, with the expectation that `collections.abc` will
come with it. This is not a guarantee.

We instead need to do an explicit `import collections.abc` along with
this.

This is a trivial change that fixes issues encountered in CI.